pyproject.toml
[tool.poetry]
name = "track-gym-shop"
version = "0.1.0"
description = ""
authors = ["EfremovAS <Alexef3@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.10.0"
pydantic-settings = "^2.3.4"
fastapi = "^0.111.1"
pytest = "^8.3.2"
httpx = "^0.27.0"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
aioredis = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
